requirements.txt
Flask
Flask-Restful
Flask-JWT
Flask-SQLAlchemy
Flask-JWT-Extended ==4.3.1
Flask-Marshmallow
marshmallow-sqlalchemy
PyJWT
uwsgi
psycopg2
boto3
Flask-Mail
marshmallow
